Backend/requirements.txt
flask==2.3.2
flask-cors==3.0.10
psycopg2-pool==1.0.0
python-dotenv==0.19.0
bcrypt==3.2.0
PyJWT==2.1.0
gunicorn==20.1.0
